+++ .storybook/stories/elements/icons/navIcon/NavIconWrapper.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { NavIcon, NavIconProps } from '../../../src/elements/navIcon/NavIcon';
+import { NavIcon, NavIconProps } from '../../../../../src/elements/icons/navIcon/NavIcon';
 
 const NavIconWrapper: React.FC<NavIconProps> = ({ icon, selected, size, title }: NavIconProps) => {
     return (
