+++ .storybook/stories/elements/icons/tagIcon/TagIconWrapper.stories.tsx
