+++ src/components/buttonWithPopover/ButtonWithPopover.tsx
@@ -1,7 +1,7 @@
 import { themeConst } from '@equinor/echo-framework';
 import { Button, CircularProgress } from '@equinor/eds-core-react';
 import React, { useState } from 'react';
-import { Icon } from '../../elements/icon/Icon';
+import { Icon } from '../../elements/icons/icon/Icon';
 import { DataInformation } from '../../types/dataInformation';
 import DataInfoPopover from '../contextMenuPopover/DataInfoPopover';
 import style from './buttonWithPopover.module.css';
