+++ src/components/tagContextMenu/TagContextMenu.tsx
@@ -1,8 +1,8 @@
 import { themeConst } from '@equinor/echo-framework';
 import { Button, Typography } from '@equinor/eds-core-react';
 import React from 'react';
-import { Icon } from '../../elements/icon/Icon';
-import { TagIconShadowWrapper } from '../../elements/tagIcon/TagIconShadow';
+import { Icon } from '../../elements/icons/icon/Icon';
+import { TagIconShadowWrapper } from '../../elements/icons/tagIcon/TagIconShadow';
 import styles from './tagContextMenu.module.css';
 
 interface TagContextMenuProps {
