+++ src/elements/icons/tagIcon/tagIcon.module.css
