+++ src/elements/icons/textIcon/TextIcon.tsx
