+++ src/elements/icons/textIcon/textIcon.module.css
