+++ src/index.ts
@@ -2,8 +2,8 @@ export { ButtonWithPopover } from './components/buttonWithPopover/ButtonWithPopo
 export { ContextMenu } from './components/contextMenu/ContextMenu';
 export { DataInfoButton } from './components/contextMenuPopover/DataInfoButton';
 export { TagContextMenu } from './components/tagContextMenu/TagContextMenu';
-export { Icon } from './elements/icon/Icon';
-export { TagIcon } from './elements/tagIcon/TagIcon';
-export { TagIconShadowWrapper } from './elements/tagIcon/TagIconShadow';
+export { Icon } from './elements/icons/icon/Icon';
+export { TagIcon } from './elements/icons/tagIcon/TagIcon';
+export { TagIconShadowWrapper } from './elements/icons/tagIcon/TagIconShadow';
 export { getIcon } from './helpers/getIcon';
 export * from './types';
